// src/user/dto/create-user.dto.ts
import { IsNotEmpty } from 'class-validator';
import { Pass } from 'src/pass/pass.entity';

export class CreateUserDTO {
  @IsNotEmpty({ message: 'username field cannot be empty' })
  username: string;
  @IsNotEmpty({ message: 'firstname field cannot be empty' })
  firstname: string;
  @IsNotEmpty({ message: 'lastname field cannot be empty' })
  lastname: string;
  @IsNotEmpty({ message: 'password field cannot be empty' })
  password: string;
  @IsNotEmpty({ message: 'email field cannot be empty' })
  email: string;
  @IsNotEmpty({ message: 'role field cannot be empty' })
  role: string;
  picture?: string;
  pass?: Pass;
}